requirements.txt
python-telegram-bot==20.3
requests
httpx==0.27.0